+++ actionview/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Add `expires_in:` option to `render` when used with `collection:`
+
+    *Jarrett Lusso*
 
 Please check [7-2-stable](https://github.com/rails/rails/blob/7-2-stable/actionview/CHANGELOG.md) for previous changes.